[DITP-pilotage/pilote-2] refactor: extraction du hook useIndicateurForm + réorg components/indicateurs
Extrait tout le câblage react-hook-form du formulaire indicateur (schéma zod
mode-dépendant, resolver, useFieldArray, useWatch, buildInitialValues,
toUpsertBody) dans un hook dédié useIndicateurForm, pour que IndicateurForm ne
porte plus que le rendu. Les deux fichiers sont regroupés dans
src/components/indicateurs/ (aligné sur l'organisation de mb-webapp). Routes
create/edit mises à jour. Pur refactor, aucun changement de comportement.

Refs:
Project: pilote-2
Client: DITP-pilotage
Category: refactor
Stack: typescript, react
Tags: mb-admin, react-hook-form, refactor
Files-changed: 4
